add smoke fetch script + fixes uncovered by first ingest
- tools/fetch_smoke.py + make fetch-smoke (4 PD videos)
- pyproject: torch index cu128 -> cu129 (cu128 lacks compute_120 PTX,
  fails NVRTC JIT on sm_121 / GB10)
- store.py: qdrant-client 1.18 dropped .search(), use .query_points().points
- caption.py, embed_video.py: torch_dtype= -> dtype=
- embed_text.py: get_sentence_embedding_dimension fallback
- docker-compose: pin qdrant 1.12.4 (1.18+ rejects on_disk segments)
- .gitignore: videos/, logs/

+++ src/ten/caption.py
@@ -71,7 +71,7 @@ class TransformersCaptioner:
             dtype = _DTYPE_MAP.get(CONFIG.dtype, torch.bfloat16)
             self._processor = AutoProcessor.from_pretrained(self.model_id)
             self._model = AutoModelForImageTextToText.from_pretrained(
-                self.model_id, torch_dtype=dtype, device_map=CONFIG.device
+                self.model_id, dtype=dtype, device_map=CONFIG.device
             )
             self._model.eval()
